% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{compute_rug}
\alias{compute_rug}
\title{compute_rug}
\usage{
compute_rug(
  dataset,
  pseudotime_column,
  group_column = "gate",
  mySeq = NULL,
  num_bins = 100,
  groups_to_keep = NULL,
  recompute_pseudotime_bins = TRUE
)
}
\description{
outputs a dataset to ggplot a discretized rug along the pseudotime based on categorical labels like cell cycle gates. Each bin is assigned a label based on a label's normalized frequency.
}
