% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{DBPN}
\alias{DBPN}
\title{DBPN}
\usage{
DBPN(
  dataset,
  density_bins = 1024,
  dataset.subset_to_use = NULL,
  column_to_normalize = NULL,
  bandwidth = "nrd0",
  adjust.value = 1,
  CIRCULARIZE = FALSE,
  RETURN_DENSITY_COORDINATES = FALSE
)
}
\description{
DBPN
}
\details{
to add:
cosine analysis
 geometric sketching
 density based normalization
}
