% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cytof_data}
\alias{example_cytof_data}
\title{My Example Dataset}
\format{
A single-cell table
...
}
\source{
The authors
}
\usage{
example_cytof_data
}
\description{
A dataset containing example data for demonstration purposes.
}
\keyword{datasets}
