% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{find_closest_neighbor_distance}
\alias{find_closest_neighbor_distance}
\title{find_closest_neighbor_distance}
\usage{
find_closest_neighbor_distance(
  full_data,
  LM_col = NULL,
  LM_group,
  features,
  k_ave = 2,
  k_solo = NULL,
  return_knn_graph = FALSE,
  return_nearest_cellid = FALSE
)
}
\description{
Calculates distance to landmark cells.
}
