% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{hafez_landmark_processing}
\alias{hafez_landmark_processing}
\title{hafez_landmark_processing}
\usage{
hafez_landmark_processing(
  train,
  full_data,
  features,
  method = "PCA",
  return_object = FALSE
)
}
\description{
Performs landmark preprocessing
}
