% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvizclust.R
\name{hafez_tsvz}
\alias{hafez_tsvz}
\title{hafez_tsvz}
\usage{
hafez_tsvz(
  ts_input,
  groups = NULL,
  approach = c("density", "expr", "other"),
  k = 5,
  dims = 2,
  type = "partitional",
  distance = "dtw_basic",
  normalize = TRUE,
  seed = 0,
  return_dDRclust = FALSE,
  ...
)
}
\description{
Performs time-series distance computation, dDR visualization, and dDR clustering. If multiple clusters_k are provided, it calculates the optimal cluster k based on minimizing intra-cluster distance. Takes as input your processed data (eg, density outputs, smoothed expression, etc). If approach is density or expr, it expects daat structure that is outputed from hafez functions to generate the density or expr values.
Otherwise, you have to generate the input data yourself, organized with the same structure. For example,column x:pseudotime that is interpolated to matcha ll timepoints, column y: the corresponding value, and columns separating your groups.
To test different clusters, you can input a vector of clusters_k. eg, c(2,3).
Cluster patterns are also outputted as a separate object
}
