% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{hafezify}
\alias{hafezify}
\title{hafezify}
\usage{
hafezify(dataset)
}
\description{
Takes as input your dataframe/tibble and outputs the same object with necessary columns to run other hafez package functions.
}
