% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{mahalanobis_distance_to_reference_group}
\alias{mahalanobis_distance_to_reference_group}
\title{mahalanobis_distance_to_reference_group}
\usage{
mahalanobis_distance_to_reference_group(
  dataset,
  features,
  reference_group_column,
  reference_group_name,
  referenceSelf = FALSE,
  sampleInternally = NULL,
  tol = 1e-20,
  precalculated_mean = NULL,
  precalculated_cov = NULL,
  CELL_COUNT_THRESHOLD = NULL
)
}
\description{
mahalanobis_distance_to_reference_group
}
