% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{calculate_cell_density_by_group}
\alias{calculate_cell_density_by_group}
\title{calculate_cell_density_by_group}
\usage{
calculate_cell_density_by_group(
  dataset,
  group_splits = NULL,
  pseudotime_column_name,
  adjust.value = 1,
  n = 512,
  from = NULL,
  to = NULL,
  COMPUTE_CIRCULAR = FALSE,
  count_threshold = 100,
  sep_operator = "_-_",
  verbose = FALSE,
  ...
)
}
\description{
calculate_cell_density_by_group
}
