% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{computePseudotimeBins}
\alias{computePseudotimeBins}
\title{computePseudotimeBins}
\usage{
computePseudotimeBins(
  dataset,
  pseudotime_column,
  interval_sequence = seq(0, 1, 0.01)
)
}
\description{
Discretizes pseudotime bins
}
