% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{hafez_DBPN}
\alias{hafez_DBPN}
\title{hafez_DBPN}
\usage{
hafez_DBPN(
  dataset,
  density_bins = 1024,
  normalize_by_sample_column = NULL,
  dataset.subset_to_use = NULL,
  column_to_normalize = NULL,
  bandwidth = "nrd0",
  adjust.value = 1,
  RETURN_DENSITY_COORDINATES = FALSE,
  new_dbp_name = "PSEUDOTIME_NORMALIZED"
)
}
\description{
Performs Density based pseudotime normalization
use the normalize_by_sample_column to specify the column name that contains the column identifier to calculate density distributions for each subject/sample of choice, then normalize the pseudotime by the density estimate
}
