% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{hafez_detect_inflections}
\alias{hafez_detect_inflections}
\title{hafez_detect_inflections}
\usage{
hafez_detect_inflections(dataset, groups = c(NULL), slope_threshold = 0.05)
}
\description{
Detects inflection points along smoothed data using a melted format dataframe (eg, output of hafez_smoothing()).
Columns must be x (pseudotime), y (fitted value), and your relevant conditions of interest that you want to stratify (eg, donor, treatment, etc).
}
