% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvizclust.R
\name{hafez_smoothing}
\alias{hafez_smoothing}
\title{hafez_smoothing}
\usage{
hafez_smoothing(
  data,
  groups = c(NULL),
  features,
  pseudotime_column = "pst",
  pst_interval = NULL,
  bs = "cs",
  ...
)
}
\description{
Computes a generalized additive model smoothing using a spline. Takes in the data, your grouping variables, and the pseudotime column name.
}
