% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{hafez_TI}
\alias{hafez_TI}
\title{hafez_TI}
\usage{
hafez_TI(
  FULL_DATA,
  features,
  features_for_start_cell_id,
  LM_DATA = NULL,
  return_pseudotime_only = TRUE,
  NumNodes = 5,
  lambda = 0.01,
  mu = 0.01,
  nReps = 30,
  ProbPoint = 0.8,
  MaxNumberOfIterations = 30,
  branch_type = c("tree", "curve", "circle"),
  TRAINING_VARIABLE = "WT",
  return_node_pos = FALSE,
  use_start_label = NULL,
  start_label_column_category = c("celltype", "HSC"),
  verbose = F
)
}
\description{
hafez_TI
}
