% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fxn.R
\name{pseudotime_mapped_mahalanobis_analysis}
\alias{pseudotime_mapped_mahalanobis_analysis}
\title{pseudotime_mapped_mahalanobis_analysis}
\usage{
pseudotime_mapped_mahalanobis_analysis(
  dataset,
  method = "landmark_mapped",
  pseudotime_bin_column = "pseudotime_bins",
  features,
  reference_group_column = NULL,
  reference_group_name = NULL,
  CELL_COUNT_THRESHOLD = NULL
)
}
\description{
pseudotime_mapped_mahalanobis_analysis
}
